requirements.txt
astroid==1.6.2
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
comet-ml==1.0.7
cycler==0.10.0
dj-database-url==0.5.0
Django==2.0.3
django-crispy-forms==1.7.2
django-markdown-deux==1.0.5
django-pagedown==1.0.4
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.8.1
idna==2.6
isort==4.3.4
Keras==2.1.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
markdown2==2.3.5
matplotlib==2.2.2
mccabe==0.6.1
mysqlclient==1.3.12
numpy==1.14.2
pandas==0.22.0
patsy==0.5.0
Pillow==5.0.0
psycopg2==2.7.4
PyJWT==1.6.1
pylint==1.8.3
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2018.3
PyYAML==3.12
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.1
six==1.11.0
sklearn==0.0
Theano==1.0.1
urllib3==1.22
websocket-client==0.47.0
whitenoise==3.3.1
wrapt==1.10.11
wurlitzer==1.0.1
